+++ kernel/c/utils/string.c
@@ -1,7 +1,7 @@
 #include "../header.h"
 
 /* this is a real one-liner. if you can manage to make it less readable, don't hesitate to do so */
-size_t strlen(const char *s) {
+unsigned strlen(const char *s) {
 	size_t len = 0; while (s[len]) len++; return len;
 }
